Move serializable `Output` into new `process` module
